Fixed to display the right icons for the channels context menu Added a warn MessageBox
git-svn-id: http://svn.code.sf.net/p/retroshare/code/trunk@7344 b45a01b8-16f6-495d-af2f-9b41ad6348cc

+++ retroshare-gui/src/gui/Identity/IdDialog.cpp
@@ -630,13 +630,24 @@ void IdDialog::removeIdentity()
 		std::cerr << std::endl;
 		return;
 	}
+	
+		QString queryWrn;
+	  queryWrn.clear();
+	  queryWrn.append(tr("Do you really want to delete this Identity ?"));
+	
+	if ((QMessageBox::question(this, tr("Really delete? "),queryWrn,QMessageBox::Yes|QMessageBox::No, QMessageBox::Yes))== QMessageBox::Yes)
+		{
+      std::string keyId = item->text(RSID_COL_KEYID).toStdString();
 
-	std::string keyId = item->text(RSID_COL_KEYID).toStdString();
+      uint32_t dummyToken = 0;
+      RsGxsIdGroup group;
+      group.mMeta.mGroupId=RsGxsGroupId(keyId);
+      rsIdentity->deleteIdentity(dummyToken, group);
+		}
+		else
+			return;
 
-	uint32_t dummyToken = 0;
-	RsGxsIdGroup group;
-	group.mMeta.mGroupId=RsGxsGroupId(keyId);
-	rsIdentity->deleteIdentity(dummyToken, group);
+
 
 }
